Restructuring intersections module for more clarity
This new module separation between object and ray representations makes more sense to me
